app/requirements.txt
Flask==2.3.3
# Deployment triggered
# Retry deployment
# Deployment successful!
# Fix EKS access
psycopg2-binary==2.9.7
